backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Django REST API for Med-Stalker (JWT auth, patients, diagnoses, prescriptions)"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=6.0.4",
    "django-cors-headers>=4.9.0",
    "djangorestframework>=3.17.1",
    "djangorestframework-simplejwt>=5.5.1",
    "faker>=37.1.0",
    "python-dotenv>=1.0.1",
]
